+++ Computer_graphics/lab_04/algorithms/bresenham.cpp
@@ -1,43 +1,11 @@
 #include "algorithms.h"
-figure_t bresenham_circle(point_t center, double radius)
+figure_t bresenham_circle(point_t center, int radius)
 {
     std::vector<pixel_t> pixels;
-
-
-    // int x = 0;
-    // int y = radius;
-    // int delta = 1 - 2 * radius;
-    // int error = 0;
-    // while(y >= 0)
-    // {
-    //     // setPixel(x0 + x, y0 + y);
-    //     // setPixel(x0 + x, y0 - y);
-    //     // setPixel(x0 - x, y0 + y);
-    //     // setPixel(x0 - x, y0 - y);
-
-    //     pixels.push_back(pixel_create(x + center.x, y + center.y));
-
-    //     error = 2 * (delta + y) - 1;
-    //     if(delta < 0 && error <= 0) {
-    //         ++x;
-    //         delta += 2 * x + 1;
-    //         continue;
-    //     }
-    //     if(delta > 0 && error > 0) {
-    //         --y;
-    //         delta += 1 - 2 * y;
-    //         continue;
-    //     }
-    //     ++x;
-    //     delta += 2 * (x - y);
-    //     --y;
-    // }
 
     int x = 0;
     int y = radius;
 
-
-    // draw_simetric_pixels(canvas, [x + xc, y + yc, colour], xc, yc, circle=True)
     pixels.push_back(pixel_create(x + center.x, y + center.y));
 
     int delta = 2 * (1 - radius);
@@ -55,7 +23,6 @@ figure_t bresenham_circle(point_t center, double radius)
         else
             delta += x + x + 1;
 
-        // draw_simetric_pixels(canvas, [x + xc, y + yc, colour], xc, yc, circle=True)
         pixels.push_back(pixel_create(x + center.x, y + center.y));
     }
 
@@ -69,8 +36,6 @@ figure_t bresenham_ellipse(point_t center, point_t radius)
     int x = 0;
     int y = radius.y;
 
-    // if (draw)
-        // draw_simetric_pixels(canvas, [x + xc, y + yc, colour], xc, yc, circle=False);
     pixels.push_back(pixel_create(x + center.x, y + center.y));
 
     int sqr_ra = radius.x * radius.x;
@@ -113,9 +78,6 @@ figure_t bresenham_ellipse(point_t center, point_t radius)
             delta += sqr_rb * (2 * x + 1) + sqr_ra * (1 - 2 * y);
         }
 
-
-        // if draw:
-        // draw_simetric_pixels(canvas, [x + xc, y + yc, colour], xc, yc, circle=False)
         pixels.push_back(pixel_create(x + center.x, y + center.y));
     }
